moved email from JWT logic to a separate file
